requirements.txt
opencv-python-headless
mediapipe
flask
flask-cors
gunicorn
numpy
supabase
requests
python-dotenv
